feat: Upgrade CLI and UI dependencies, streamline code, and enhance async disposal
This upgrade introduces a revamped dependency structure for the CLI,
eliminating certain dependencies and adding others like `mime-types`,
`nanostores`, and `@types/mime-types`, reflecting a refined purpose and
scope of the module. Simplifications to async code patterns have been
made alongside corrections to logic flow, particularly in the main CLI
logic and `execCmd` function, to ensure that the help message is
displayed as expected. The UI component `render-frame` has been retooled
with a dynamic static asset handler and simplified server setup. Removal
of the obsolete `run-astro.ts` is a part of cleanup activities as the
project moves away from previous setups. A new utility for resource
disposal, `createContext` and `createContextAsync`, has been introduced
and is underlined by the inclusion of corresponding tests. These changes
collectively aim at better maintainability and a leaner stack.