docs: standardize job input parameter structure in README code examples
In this commit, the code examples within the README file have been updated to follow a consistent structure for job input parameters. Previously, the job functions 'write file' and 'get number' were directly deconstructing their parameters from the arguments object. This has been modified to deconstruct the parameters from an `input` object instead, which wraps the original parameters. This change aims to standardize the way parameters are passed to job functions, making the code more uniform and easier to understand for developers. This modification does not affect the underlying functionality but makes the code examples within the README more consistent and aligned with best practices for object destructuring in modern JavaScript.