chore(accion-common): bump version to 0.2.1
This commit updates the version of `accion-common` from `0.2.0` to `0.2.1`. The increment aligns with semantic versioning principles, suggesting that the changes include minor fixes or non-breaking enhancements. Besides ensuring the package remains up to date, this version bump is also crucial for making sure dependent projects will be able to pull the latest changes without compatibility issues. This step is part of the routine maintenance or improvement of the library, possibly in response to community feedback, bug reports, or the addition of small features.