chore(accion-common): add module type and bump version to 0.2.3
This commit introduces a significant change to the `@accions/common` package by specifying its type as a "module" in the `package.json`. This alteration is pivotal for supporting ES Module (ESM) notation, enhancing the package's compatibility with modern JavaScript projects that leverage import/export syntax. Additionally, the version has been incremented from 0.2.2 to 0.2.3 to reflect this enhancement and comply with semantic versioning practices. This version bump signals to users that a new feature has been added, encouraging them to update their packages to utilize the latest capabilities. Implementing ESM support is a strategic move to ensure the package remains relevant and functional within the evolving JavaScript ecosystem.