chore(accion-common): bump package version to 0.2.4
This commit increments the version of the `accion-common` package from 0.2.3 to 0.2.4. This version bump is necessary to prepare for the release of new features, improvements, or bug fixes introduced in this cycle. Remember, consuming applications need to update their dependencies to take advantage of the changes. This change is part of our ongoing effort to maintain and improve the quality of our packages through iterative updates.